Update package-lock.json
The dotenv package was missing here, it's good to keep the lockfile up
to date it so folks can use `npm ci` for installs and get identical
versions as the lockfile.

The license field also appears to have been wrong in package-lock.json.
No idea how that would have happened since it appears to have been MIT
in package.json since the first commit, but `npm i` fixed it so I'm
committing it.