Fix rate limiting
Axios's default behavior for 429 responses is to reject the promise
(which raises an exception with `await`). See docs of `validateStatus`
and the default value at
https://github.com/axios/axios/blob/5064b108de336ff34862650709761b8a96d26be0/lib/defaults/index.js#L145.
The default behavior is to reject/raise for anything outside the 200
range.

This patch tweaks the axios config to consider 200 & 429 responses valid
so the rate limiting behavior can work as expected.